changed int priority to float

+++ com.unity.shadergraph/Editor/Drawing/Blackboard/BlackboardInputInfo.cs
@@ -5,7 +5,7 @@ namespace UnityEditor.ShaderGraph
     [AttributeUsage(AttributeTargets.Class)]
     class BlackboardInputInfo : Attribute
     {
-        public int priority;
+        public float priority;
         public string name;
 
         /// <summary>
@@ -13,7 +13,7 @@ namespace UnityEditor.ShaderGraph
         /// </summary>
         /// <param name="priority">Priority of the item, higher values will result in lower positions in the menu.</param>
         /// <param name="name">Name of the item. If null, the class name of the item will be used instead.</param>
-        public BlackboardInputInfo(int priority, string name = null)
+        public BlackboardInputInfo(float priority, string name = null)
         {
             this.priority = priority;
             this.name = name;
